Authentication and Authorization

+++ controllers/api/v1/posts_api.js
@@ -26,13 +26,20 @@ module.exports.destroy=async function(req,res){
             //post.user contains id of the user who posted that post.
             //req.user contains info of current logged in user.
             // .id means converting the object id into string.
+            if(post.user==req.user.id){
           
                 post.remove();
            
                 await Comment.deleteMany({post:req.params.id});
                 return res.json(200,{
                     message:'Post and associated comments deleted!!'
-                })
+                });
+            }
+            else{
+                return res.json(401, {
+                    message: "You cannot delete this post!"
+                });
+            }
         }
     catch(err){
        console.log('********error',err);
